requirements.txt
appdirs==1.4.4
certifi==2021.5.30
charset-normalizer==2.0.4
cityscapesScripts==2.2.0
coloredlogs==15.0.1
cycler==0.10.0
Cython==0.29.24
filelock==3.0.12
fvcore==0.1.5.post20211023
gdown==3.13.0
humanfriendly==9.2
idna==3.2
iopath==0.1.9
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.2
packaging==21.0
Pillow==8.0.1
pip-check-reqs==2.3.0
pip-chill==1.0.1
portalocker==2.3.2
ptflops==0.6.6
pycocotools==2.0.2
pyparsing==2.4.7
PyQt5==5.15.4
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyquaternion==0.9.9
PySocks==1.7.1
python-dateutil==2.8.2
PyYAML==5.4.1
requests==2.26.0
scipy==1.7.1
six==1.16.0
tabulate==0.8.9
termcolor==1.1.0
timm==0.4.12
torch==1.9.0
torchvision==0.10.0
tqdm==4.62.2
typing==3.7.4.3
typing-extensions==3.10.0.2
urllib3==1.26.6
yacs==0.1.8
